+++ src/main/java/io/github/thebusybiscuit/slimefun4/implementation/items/weapons/SwordOfBeheading.java
@@ -3,8 +3,11 @@ package io.github.thebusybiscuit.slimefun4.implementation.items.weapons;
 import java.util.Random;
 import java.util.concurrent.ThreadLocalRandom;
 
+import javax.annotation.ParametersAreNonnullByDefault;
+
 import org.bukkit.Material;
 import org.bukkit.entity.Creeper;
+import org.bukkit.entity.Monster;
 import org.bukkit.entity.Player;
 import org.bukkit.entity.Skeleton;
 import org.bukkit.entity.WitherSkeleton;
@@ -21,6 +24,17 @@ import me.mrCookieSlime.Slimefun.Lists.RecipeType;
 import me.mrCookieSlime.Slimefun.Objects.Category;
 import me.mrCookieSlime.Slimefun.api.SlimefunItemStack;
 
+/**
+ * The {@link SwordOfBeheading} is a special kind of sword which allows you to obtain
+ * {@link Zombie}, {@link Skeleton} and {@link Creeper} skulls when killing the respective {@link Monster}.
+ * Additionally, you can also obtain the head of a {@link Player} by killing them too.
+ * This sword also allows you to have a higher chance of getting the skull of a {@link WitherSkeleton} too.
+ * 
+ * All chances are managed by an {@link ItemSetting} and can be configured.
+ * 
+ * @author TheBusyBiscuit
+ *
+ */
 public class SwordOfBeheading extends SimpleSlimefunItem<EntityKillHandler> {
 
     private final ItemSetting<Integer> chanceZombie = new IntRangeSetting("chance.ZOMBIE", 0, 40, 100);
@@ -29,6 +43,7 @@ public class SwordOfBeheading extends SimpleSlimefunItem<EntityKillHandler> {
     private final ItemSetting<Integer> chanceCreeper = new IntRangeSetting("chance.CREEPER", 0, 40, 100);
     private final ItemSetting<Integer> chancePlayer = new IntRangeSetting("chance.PLAYER", 0, 70, 100);
 
+    @ParametersAreNonnullByDefault
     public SwordOfBeheading(Category category, SlimefunItemStack item, RecipeType recipeType, ItemStack[] recipe) {
         super(category, item, recipeType, recipe);
 
@@ -40,29 +55,40 @@ public class SwordOfBeheading extends SimpleSlimefunItem<EntityKillHandler> {
         return (e, entity, killer, item) -> {
             Random random = ThreadLocalRandom.current();
 
-            if (e.getEntity() instanceof Zombie) {
+            switch (e.getEntityType()) {
+            case ZOMBIE:
                 if (random.nextInt(100) < chanceZombie.getValue()) {
                     e.getDrops().add(new ItemStack(Material.ZOMBIE_HEAD));
                 }
-            } else if (e.getEntity() instanceof WitherSkeleton) {
-                if (random.nextInt(100) < chanceWitherSkeleton.getValue()) {
-                    e.getDrops().add(new ItemStack(Material.WITHER_SKELETON_SKULL));
-                }
-            } else if (e.getEntity() instanceof Skeleton) {
+                break;
+            case SKELETON:
                 if (random.nextInt(100) < chanceSkeleton.getValue()) {
                     e.getDrops().add(new ItemStack(Material.SKELETON_SKULL));
                 }
-            } else if (e.getEntity() instanceof Creeper) {
+                break;
+            case CREEPER:
                 if (random.nextInt(100) < chanceCreeper.getValue()) {
                     e.getDrops().add(new ItemStack(Material.CREEPER_HEAD));
                 }
-            } else if (e.getEntity() instanceof Player && random.nextInt(100) < chancePlayer.getValue()) {
-                ItemStack skull = new ItemStack(Material.PLAYER_HEAD);
-                ItemMeta meta = skull.getItemMeta();
-                ((SkullMeta) meta).setOwningPlayer((Player) e.getEntity());
-                skull.setItemMeta(meta);
+                break;
+            case WITHER_SKELETON:
+                if (random.nextInt(100) < chanceWitherSkeleton.getValue()) {
+                    e.getDrops().add(new ItemStack(Material.WITHER_SKELETON_SKULL));
+                }
+                break;
+            case PLAYER:
+                if (random.nextInt(100) < chancePlayer.getValue()) {
+                    ItemStack skull = new ItemStack(Material.PLAYER_HEAD);
 
-                e.getDrops().add(skull);
+                    ItemMeta meta = skull.getItemMeta();
+                    ((SkullMeta) meta).setOwningPlayer((Player) e.getEntity());
+                    skull.setItemMeta(meta);
+
+                    e.getDrops().add(skull);
+                }
+                break;
+            default:
+                break;
             }
         };
     }