+++ CHANGELOG.md
@@ -72,6 +72,8 @@
 * Fixed #2558
 * Fixed a duplication bug with the Block Placer
 * Fixed Slimefun Guide Settings showing "last activity" as a negative number
+* Fixed Armor Stands getting damaged/pushed by Explosive Bow
+* Fixed Sword of Beheading dropping Zombie/Skeleton Skulls from Zombie/Skeleton subvariants
 
 ## Release Candidate 17 (17 Oct 2020)